Move shoot specific checks to `health`
